+++ lib/adhara.dart
@@ -5,7 +5,6 @@ export 'package:adhara/datainterface/network/network_provider.dart';
 export 'package:adhara/datainterface/storage/storage_classes.dart';
 export 'package:adhara/datainterface/storage/storage_provider.dart';
 export 'package:adhara/datainterface/storage/bean_storage_provider.dart';
-export 'package:adhara/datainterface/storage/http_storage_provider.dart';
 export 'package:adhara/datainterface/data_interface.dart';
 export 'package:adhara/resources/ri.dart';
 export 'package:adhara/resources/r.dart';